requirements.txt
asgiref==3.5.1
autopep8==1.6.0
black==22.3.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
Django==4.0.4
django-silk==4.3.0
djangorestframework==3.13.1
Faker==13.7.0
gprof2dot==2021.2.21
idna==3.3
Jinja2==3.1.2
MarkupSafe==2.1.1
model-bakery==1.5.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
pycodestyle==2.8.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
urllib3==1.26.9
